Added the profile command. Cleaned up the code and closed associated issues.  (#14)
* began work on /update command. build in logic to accept a single username and an optional [share] tag to return the

* added a new item to the update slash command. updates the user's picture
updated the schema to hold two forms of the picture (small and medium)
working correctly
need to update the help and instructions

* created the general profile card for users
called via /profile @userName
returns profile card populated with available data
currently /works/ on Slack message builder and slash command testing but the formatting is off

* profiles are returning correctly and accept a share parameter
need to create the return specific item method and attachments

* fixed the promise issue with saving and responding in the processCheckin method
added valSubmit method to valueStringer.js and implemented it in the submit responses
updated the skillSubmit interaction to process a submitted valueObject or reset to the beginning of the form

* added and tested the ability to call any profile item [except projects, skills, certifications for now]
and be able to either get them returned ephemerally by default or pass share for sharing in channel. story will always default to ephemeral to prevent
spamming

* fixed formatting on profile card

* added a certifications /profile option

* /profile no longer requires a hashtag prefix to the item

* added profile response for projects and certifications sub options

* added missing .gitignore file. restructured projectsItemResponse to use alternating colors with alternateAttachmentColor function

* removing local packge-lock.json

* fixed package-lock typo in .gitignore

* removing package lock

* removed .idea

* finished the /profile command

* minor changes to prepare for merging.

* solved issue #12: updating skills. tested and working bug-free

* fixed processUpdate database method to correctly function as a promise. allows for more informative responses
based on the success or failure of an actual database save

* added a help section for updating the profile picture

* fixed the help text for updating

* fixed issue #11:
did not rewrite the error scanner. it functions as needed - can consider refactoring in the future but right now it is catching and responding to all errors correctly.
now correctly catches project updates that are missing a gitHub repo link

* added help responses to close issue #9